+++ server.js
@@ -15,7 +15,6 @@ if (process.env.NODE_ENV !== 'production') {
 }
 
 app.use('/forms', require('./routes/form.routes.js'));
-app.use('/send', require('./routes/mailer.routes.js'));
 app.use('/apartments', require('./routes/apartment.routes.js'));
 
 // set port, listen for requests
